backend/requirements.txt
fastapi
uvicorn
openai-whisper
python-multipart
PyMuPDF
rapidfuzz